Added InputStream and Terminal
InputStream can be used to read raw bytes from a input
file descriptor. For now it provides a method that reads
bytes into a vector.
Terminal is used to manipulate and restore the termios
settings of the terminal. This is useful for enabling
raw mode. When other features are deemed useful, they
will be added.