Added support for non-blocking reads
When the input file descriptor has the O_NONBLOCK flag set, the
call to read may fail with a result of -1 setting errno to either
EAGAIN or EWOULDBLOCK. Note that EAGAIN and EWOULDBLOCK may have
the same value, but that is not necessary. This commit prevents
InputStream from throwing an exception if the file descriptor has
O_NONBLOCK set.