Created stub class for screen control
I am still trying to figure out the purpose of the Screen class,
but it will probably abstract most of the interaction with the
terminal/screen. Other editors have used client/server architectures
to allow multiple interactions with a single buffer. I was not
planning to do this, but it might allow the interaction to be
further abstracted. If ed had its own protocol for user interactions
and text movement, that would define a clear API for different
implementations.